+++ server/test/integration/com/thoughtworks/go/server/service/GoConfigServiceIntegrationTest.java
@@ -35,7 +35,6 @@ import com.thoughtworks.go.server.dao.DatabaseAccessHelper;
 import com.thoughtworks.go.server.domain.Username;
 import com.thoughtworks.go.server.service.result.HttpLocalizedOperationResult;
 import com.thoughtworks.go.server.service.result.LocalizedOperationResult;
-import com.thoughtworks.go.service.ConfigRepository;
 import com.thoughtworks.go.util.GoConfigFileHelper;
 import com.thoughtworks.go.util.ReflectionUtil;
 import org.junit.After;
@@ -48,7 +47,6 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.util.Arrays;
 
 import static javax.servlet.http.HttpServletResponse.SC_CONFLICT;
 import static org.hamcrest.Matchers.containsString;
@@ -70,9 +68,7 @@ public class GoConfigServiceIntegrationTest {
     @Autowired private GoConfigService goConfigService;
     @Autowired private DatabaseAccessHelper dbHelper;
     @Autowired private Localizer localizer;
-    @Autowired private ConfigRepository configRepo;
     @Autowired private CachedGoConfig cachedGoConfig;
-    @Autowired private ServerConfigService serverConfigService;
 
     private GoConfigFileHelper configHelper;
 
@@ -788,15 +784,6 @@ public class GoConfigServiceIntegrationTest {
         GoConfigValidity validity = saver.saveXml(os.toString(), user2SeeingMd5);
 
         assertThat(validity.isValid(), is(true));
-    }
-
-    @Test
-    public void shouldUpdateExistingPipelineSelectionsForAnonymousUserIfItAlreadyExistsInsteadOfAddingDuplicates() {
-        long firstSaveId = goConfigService.persistSelectedPipelines(null, Arrays.asList("pipeline1", "pipeline2"), false);
-        long secondSaveId = goConfigService.persistSelectedPipelines(null, Arrays.asList("pipeline1"), false);
-
-        assertThat(firstSaveId, is(secondSaveId));
-        assertThat(goConfigService.getSelectedPipelines(null).getSelections(), is("pipeline1"));
     }
 
     private void setJobTimeoutTo(final String jobTimeout) {
