+++ server/test/unit/com/thoughtworks/go/server/service/GoConfigServiceTest.java
@@ -633,149 +633,6 @@ public class GoConfigServiceTest {
     }
 
     @Test
-    public void shouldPersistPipelineSelections_WhenSecurityIsDisabled() {
-        Date date = new DateTime(2000, 1, 1, 1, 1, 1, 1).toDate();
-        when(clock.currentTime()).thenReturn(date);
-        mockConfig();
-        Matcher<PipelineSelections> pipelineSelectionsMatcher = hasValues(Arrays.asList("pipelineX", "pipeline3"), Arrays.asList("pipeline1", "pipeline2"), date, null);
-        when(pipelineRepository.saveSelectedPipelines(argThat(pipelineSelectionsMatcher))).thenReturn(2L);
-        assertThat(goConfigService.persistSelectedPipelines(null, Arrays.asList("pipelineX", "pipeline3"), true), is(2l));
-        verify(pipelineRepository).saveSelectedPipelines(argThat(pipelineSelectionsMatcher));
-    }
-
-    @Test
-    public void shouldPersistPipelineSelectionsAgainstUser_AlreadyHavingSelections() {
-        Date date = new DateTime(2000, 1, 1, 1, 1, 1, 1).toDate();
-        when(clock.currentTime()).thenReturn(date);
-        mockConfigWithSecurity();
-        User user = getUser("badger", 10L);
-        PipelineSelections pipelineSelections = new PipelineSelections(Arrays.asList("pipeline2"), new Date(), user.getId(), true);
-        when(pipelineRepository.findPipelineSelectionsByUserId(user.getId())).thenReturn(pipelineSelections);
-        when(pipelineRepository.saveSelectedPipelines(pipelineSelections)).thenReturn(2L);
-
-        long pipelineSelectionId = goConfigService.persistSelectedPipelines(user.getId(), Arrays.asList("pipelineX", "pipeline3"), true);
-
-        assertThat(pipelineSelections.getSelections(), is("pipeline1,pipeline2"));
-        assertThat(pipelineSelectionId, is(2l));
-        verify(pipelineRepository).saveSelectedPipelines(pipelineSelections);
-        verify(pipelineRepository).findPipelineSelectionsByUserId(user.getId());
-    }
-
-    @Test
-    public void shouldPersistPipelineSelectionsAgainstUser_WhenUserHasNoSelections() {
-        Date date = new DateTime(2000, 1, 1, 1, 1, 1, 1).toDate();
-        when(clock.currentTime()).thenReturn(date);
-        mockConfigWithSecurity();
-        User user = getUser("badger", 10L);
-        Matcher<PipelineSelections> pipelineSelectionsMatcher = hasValues(Arrays.asList("pipelineX", "pipeline3"), Arrays.asList("pipeline1", "pipeline2"), date, user.getId());
-        when(pipelineRepository.findPipelineSelectionsByUserId(user.getId())).thenReturn(null);
-        when(pipelineRepository.saveSelectedPipelines(argThat(pipelineSelectionsMatcher))).thenReturn(2L);
-
-        long pipelineSelectionsId = goConfigService.persistSelectedPipelines(user.getId(), Arrays.asList("pipelineX", "pipeline3"), true);
-
-        assertThat(pipelineSelectionsId, is(2l));
-        verify(pipelineRepository).saveSelectedPipelines(argThat(pipelineSelectionsMatcher));
-        verify(pipelineRepository).findPipelineSelectionsByUserId(user.getId());
-    }
-
-    @Test
-    public void shouldPersistPipelineSelectionsShouldRemovePipelinesFromSelectedGroups() {
-        CruiseConfig config = configWith(
-                createGroup("group1", pipelineConfig("pipeline1"), pipelineConfig("pipeline2")),
-                createGroup("group2", pipelineConfig("pipelineX")),
-                createGroup("group3", pipelineConfig("pipeline3"), pipelineConfig("pipeline4")));
-        when(goConfigDao.load()).thenReturn(config);
-        goConfigService.persistSelectedPipelines(null, Arrays.asList("pipeline1", "pipeline2", "pipeline3"), true);
-        verify(pipelineRepository).saveSelectedPipelines(argThat(hasValues(Arrays.asList("pipeline1", "pipeline2", "pipeline3"), Arrays.asList("pipelineX", "pipeline4"), clock.currentTime(), null)));
-    }
-
-    @Test
-    public void shouldPersistInvertedListOfPipelineSelections_WhenBlacklistIsSelected() {
-        Date date = new DateTime(2000, 1, 1, 1, 1, 1, 1).toDate();
-        when(clock.currentTime()).thenReturn(date);
-        mockConfigWithSecurity();
-
-        User user = getUser("badger", 10L);
-        PipelineSelections blacklistPipelineSelections = new PipelineSelections(new ArrayList<String>(), date, user.getId(), false);
-        when(pipelineRepository.findPipelineSelectionsByUserId(user.getId())).thenReturn(blacklistPipelineSelections);
-
-        goConfigService.persistSelectedPipelines(user.getId(), Arrays.asList("pipelineX", "pipeline3"), true);
-
-        verify(pipelineRepository).saveSelectedPipelines(argThat(isAPipelineSelectionsInstanceWith(true, "pipeline1", "pipeline2")));
-    }
-
-    @Test
-    public void shouldPersistNonInvertedListOfPipelineSelections_WhenWhitelistIsSelected() {
-        Date date = new DateTime(2000, 1, 1, 1, 1, 1, 1).toDate();
-        when(clock.currentTime()).thenReturn(date);
-        mockConfigWithSecurity();
-
-        User user = getUser("badger", 10L);
-        PipelineSelections whitelistPipelineSelections = new PipelineSelections(new ArrayList<String>(), date, user.getId(), true);
-        when(pipelineRepository.findPipelineSelectionsByUserId(user.getId())).thenReturn(whitelistPipelineSelections);
-
-        goConfigService.persistSelectedPipelines(user.getId(), Arrays.asList("pipelineX", "pipeline3"), false);
-
-        verify(pipelineRepository).saveSelectedPipelines(argThat(isAPipelineSelectionsInstanceWith(false, "pipelineX", "pipeline3")));
-    }
-
-    @Test
-    public void shouldUpdateAlreadyPersistedSelection_WhenSecurityIsDisabled() {
-        Date date = new DateTime(2000, 1, 1, 1, 1, 1, 1).toDate();
-        when(clock.currentTime()).thenReturn(date);
-        mockConfig();
-        PipelineSelections pipelineSelections = new PipelineSelections(Arrays.asList("pip1"));
-        when(pipelineRepository.findPipelineSelectionsByUserId(null)).thenReturn(pipelineSelections);
-        List<String> newPipelines = Arrays.asList("pipeline1", "pipeline2");
-
-        goConfigService.persistSelectedPipelines(null, newPipelines, true);
-
-        assertHasSelected(pipelineSelections, newPipelines);
-        assertThat(pipelineSelections.lastUpdated(), is(date));
-        verify(pipelineRepository).findPipelineSelectionsByUserId(null);
-        verify(pipelineRepository).saveSelectedPipelines(argThat(hasValues(Arrays.asList("pipeline1", "pipeline2"), Arrays.asList("pipelineX", "pipeline3"), clock.currentTime(), null)));
-    }
-
-    @Test
-    public void shouldReturnLatestPersistedPipelineSelectionsAgainstNullUser_WhenSecurityisDisabled() {
-        PipelineSelections pipelineSelections = new PipelineSelections(Arrays.asList("pip1"));
-        when(pipelineRepository.findPipelineSelectionsByUserId(null)).thenReturn(pipelineSelections);
-        assertThat(goConfigService.getSelectedPipelines(null), is(pipelineSelections));
-    }
-
-    @Test
-    public void shouldReturnAllPipelinesAgainstNullUser_WhenSecurityisDisabledAndNoSelectionsArePersistedForNullUser() {
-        when(pipelineRepository.findPipelineSelectionsByUserId(null)).thenReturn(null);
-        assertThat(goConfigService.getSelectedPipelines(null), is(PipelineSelections.ALL));
-    }
-
-    @Test
-    public void shouldReturnPersistedPipelineSelectionsAgainstUser_WhenSecurityIsEnabled() {
-        User loser = getUser("loser", 10L);
-        User newUser = getUser("new user", 20L);
-        when(userDao.findUser("new user")).thenReturn(newUser);
-        mockConfigWithSecurity();
-        PipelineSelections pipelineSelections = new PipelineSelections(Arrays.asList("pip1"));
-
-        when(pipelineRepository.findPipelineSelectionsByUserId(loser.getId())).thenReturn(pipelineSelections);
-
-        assertThat(goConfigService.getSelectedPipelines(loser.getId()), is(pipelineSelections));
-        assertThat(goConfigService.getSelectedPipelines(newUser.getId()), is(PipelineSelections.ALL));
-    }
-
-    @Test
-    public void shouldReturnAllPipelineSelections_WhenSecurityIsEnabled_AndNoPersistedSelections() {
-        User user = getUser("loser", 10L);
-        User newUser = getUser("new user", 20L);
-        when(userDao.findUser("new user")).thenReturn(newUser);
-        mockConfigWithSecurity();
-
-        when(pipelineRepository.findPipelineSelectionsByUserId(user.getId())).thenReturn(null);
-
-        assertThat(goConfigService.getSelectedPipelines(newUser.getId()), is(PipelineSelections.ALL));
-    }
-
-    @Test
     public void shouldRegisterBaseUrlChangeListener() throws Exception {
         CruiseConfig cruiseConfig = new GoConfigMother().cruiseConfigWithOnePipelineGroup();
         stub(goConfigDao.load()).toReturn(cruiseConfig);
@@ -1150,58 +1007,6 @@ public class GoConfigServiceTest {
 	}
 
     @Test
-    public void shouldNotUpdatePipelineSelectionsWhenTheUserIsAnonymousAndHasNeverSelectedPipelines() {
-        goConfigService.updateUserPipelineSelections(null, new CaseInsensitiveString("pipelineNew"));
-
-        verify(pipelineRepository, times(0)).saveSelectedPipelines(argThat(Matchers.any(PipelineSelections.class)));
-    }
-
-    @Test
-    public void shouldNotUpdatePipelineSelectionsWhenTheUserIsAnonymousAndHasSelectedPipelines_WithBlacklist() {
-        when(pipelineRepository.findPipelineSelectionsByUserId(null)).thenReturn(new PipelineSelections(Arrays.asList("pipeline1", "pipeline2"), null, null, true));
-
-        goConfigService.updateUserPipelineSelections(null, new CaseInsensitiveString("pipelineNew"));
-
-        verify(pipelineRepository).findPipelineSelectionsByUserId(null);
-        verify(pipelineRepository, times(0)).saveSelectedPipelines(argThat(Matchers.any(PipelineSelections.class)));
-    }
-
-    @Test
-    public void shouldUpdatePipelineSelectionsWhenTheUserIsAnonymousAndHasSelectedPipelines_WithWhitelist() {
-        when(pipelineRepository.findPipelineSelectionsByUserId(null)).thenReturn(new PipelineSelections(Arrays.asList("pipeline1", "pipeline2"), null, null, false));
-
-        goConfigService.updateUserPipelineSelections(null, new CaseInsensitiveString("pipelineNew"));
-
-        verify(pipelineRepository).findPipelineSelectionsByUserId(null);
-        verify(pipelineRepository, times(1)).saveSelectedPipelines(argThat(isAPipelineSelectionsInstanceWith(false, "pipeline1", "pipeline2", "pipelineNew")));
-    }
-
-    @Test
-    public void shouldNotUpdatePipelineSelectionsWhenTheUserIsLoggedIn_WithBlacklist() {
-        mockConfigWithSecurity();
-
-        when(pipelineRepository.findPipelineSelectionsByUserId(1L)).thenReturn(new PipelineSelections(Arrays.asList("pipeline1", "pipeline2"), null, null, true));
-
-        goConfigService.updateUserPipelineSelections(1L, new CaseInsensitiveString("pipelineNew"));
-
-        verify(pipelineRepository).findPipelineSelectionsByUserId(1L);
-        verify(pipelineRepository, times(0)).saveSelectedPipelines(argThat(Matchers.any(PipelineSelections.class)));
-    }
-
-    @Test
-    public void shouldUpdatePipelineSelectionsWhenTheUserIsLoggedIn_WithWhitelist() {
-        mockConfigWithSecurity();
-
-        when(pipelineRepository.findPipelineSelectionsByUserId(1L)).thenReturn(new PipelineSelections(Arrays.asList("pipeline1", "pipeline2"), null, null, false));
-
-        goConfigService.updateUserPipelineSelections(1L, new CaseInsensitiveString("pipelineNew"));
-
-        verify(pipelineRepository).findPipelineSelectionsByUserId(1L);
-        verify(pipelineRepository, times(1)).saveSelectedPipelines(argThat(isAPipelineSelectionsInstanceWith(false, "pipeline1", "pipeline2", "pipelineNew")));
-    }
-
-
-    @Test
     public void pipelineEditableViaUI_shouldReturnFalseWhenPipelineIsRemote() throws Exception
     {
         PipelineConfigs group = new BasicPipelineConfigs();
@@ -1285,21 +1090,6 @@ public class GoConfigServiceTest {
                 ConfigElementImplementationRegistryMother.withNoPlugins(), instanceFactory, null, null);
     }
 
-    private CruiseConfig mockConfig() {
-        CruiseConfig config = configWith(
-                createGroup("group0", pipelineConfig("pipeline1"), pipelineConfig("pipeline2")),
-                createGroup("group1", pipelineConfig("pipelineX")),
-                createGroup("group2", pipelineConfig("pipeline3")));
-        when(goConfigDao.load()).thenReturn(config);
-        return config;
-    }
-
-    private CruiseConfig mockConfigWithSecurity() {
-        CruiseConfig config = mockConfig();
-        config.server().useSecurity(new SecurityConfig(null, new PasswordFileConfig("path"), true));
-        return config;
-    }
-
     private GoConfigInvalidException getGoConfigInvalidException() {
         ConfigErrors configErrors = new ConfigErrors();
         configErrors.add("command", "command cannot be empty");
@@ -1342,58 +1132,6 @@ public class GoConfigServiceTest {
             public void describeTo(Description description) {
             }
         };
-    }
-
-    private User getUser(final String userName, long id) {
-        long userId = id;
-        User user = new User(userName);
-        user.setId(userId);
-        when(userDao.findUser(userName)).thenReturn(user);
-        return user;
-    }
-
-    private Matcher<PipelineSelections> hasValues(final List<String> isVisible, final List<String> isNotVisible, final Date today, final Long userId) {
-        return new BaseMatcher<PipelineSelections>() {
-            public boolean matches(Object o) {
-                PipelineSelections pipelineSelections = (PipelineSelections) o;
-                assertHasSelected(pipelineSelections, isVisible);
-                assertHasSelected(pipelineSelections, isNotVisible, false);
-                assertThat(pipelineSelections.lastUpdated(), is(today));
-                assertThat(pipelineSelections.userId(), is(userId));
-                return true;
-            }
-
-            public void describeTo(Description description) {
-            }
-        };
-    }
-
-    private Matcher<PipelineSelections> isAPipelineSelectionsInstanceWith(final boolean isBlacklist, final String... pipelineSelectionsInInstance) {
-        return new BaseMatcher<PipelineSelections>() {
-            public boolean matches(Object o) {
-                PipelineSelections pipelineSelections = (PipelineSelections) o;
-                assertThat(pipelineSelections.isBlacklist(), is(isBlacklist));
-
-                List<String> expectedSelectionsAsList = Arrays.asList(pipelineSelectionsInInstance);
-                assertEquals(pipelineSelections.getSelections(), ListUtil.join(expectedSelectionsAsList, ","));
-
-                return true;
-            }
-
-            public void describeTo(Description description) {
-            }
-        };
-    }
-
-    private void assertHasSelected(PipelineSelections pipelineSelections, List<String> pipelines) {
-        assertHasSelected(pipelineSelections, pipelines, true);
-    }
-
-    private void assertHasSelected(PipelineSelections pipelineSelections, List<String> pipelines, boolean has) {
-        String message = "Expected: " + pipelines + " to include " + pipelineSelections + ": (" + has + ").";
-        for (String pipeline : pipelines) {
-            assertThat(message + ". Failed to find: " + pipeline, pipelineSelections.includesPipeline(pipelineConfig(pipeline)), is(has));
-        }
     }
 
     private String groupXml(final String groupName) {
