+++ server/webapp/WEB-INF/rails.new/lib/services.rb
@@ -33,7 +33,7 @@ module Services
            :task_view_service, :server_status_service, :pipeline_configs_service, :pipeline_service, :material_update_service,
            :system_service, :default_plugin_manager, :command_repository_service, :value_stream_map_service, :admin_service, :config_repository, :package_repository_service, :package_definition_service, :pipeline_sql_map_dao, :pluggable_task_service, :garage_service,
            :material_config_service, :feature_toggle_service, :cc_tray_service, :pluggable_scm_service, :cc_tray_status_service, :plugin_service, :artifacts_dir_holder, :version_info_service, :entity_hashing_service,
-           :elastic_profile_service, :password_deserializer
+           :elastic_profile_service, :password_deserializer, :pipeline_selections_service
   )
 
   service_with_alias_name(:go_config_service_for_url, "goConfigService")
