+++ server/webapp/WEB-INF/rails.new/spec/controllers/admin/pipelines_controller_spec.rb
@@ -27,6 +27,7 @@ describe Admin::PipelinesController do
   before(:each) do
     @user = Username.new(CaseInsensitiveString.new("loser"))
     controller.stub(:current_user).and_return(@user)
+    controller.stub(:pipeline_selections_service).and_return(@pipeline_selections_service = double('pipeline selections service'))
   end
 
   describe "routes" do
@@ -402,7 +403,7 @@ describe Admin::PipelinesController do
 
       controller.stub(:pluggable_task_service).and_return(@pluggable_task_service)
 
-      allow(@go_config_service).to receive(:updateUserPipelineSelections)
+      allow(@pipeline_selections_service).to receive(:updateUserPipelineSelections)
       allow(controller).to receive(:cookies).and_return({})
     end
 
@@ -446,7 +447,7 @@ describe Admin::PipelinesController do
 
       @go_config_service.should_receive(:getCurrentConfig).twice.and_return(Cloner.new().deepClone(@cruise_config))
       @pipeline_pause_service.should_receive(:pause).with("new-pip", "Under construction", @user)
-      @go_config_service.should_receive(:updateUserPipelineSelections).with(current_user_entity_id, CaseInsensitiveString.new(pipeline_name))
+      @pipeline_selections_service.should_receive(:updateUserPipelineSelections).with(current_user_entity_id, CaseInsensitiveString.new(pipeline_name))
 
       stub_save_for_success
       post :create, :config_md5 => "1234abcd", :pipeline_group => {:group => "new-group", :pipeline => {:name => pipeline_name}}
@@ -456,7 +457,7 @@ describe Admin::PipelinesController do
       pipeline_name = "new-pip"
 
       @go_config_service.should_receive(:getCurrentConfig).twice.and_return(Cloner.new().deepClone(@cruise_config))
-      @go_config_service.should_not_receive(:updateUserPipelineSelections)
+      @pipeline_selections_service.should_not_receive(:updateUserPipelineSelections)
 
       stub_save_for_validation_error do |result, _, _|
         result.unauthorized(com.thoughtworks.go.i18n.LocalizedMessage.string("UNAUTHORIZED_TO_CREATE_PIPELINE"), nil)
@@ -715,7 +716,7 @@ describe Admin::PipelinesController do
     describe "save_clone" do
       before :each do
         allow(@pipeline_pause_service).to receive(:pause)
-        allow(@go_config_service).to receive(:updateUserPipelineSelections)
+        allow(@pipeline_selections_service).to receive(:updateUserPipelineSelections)
       end
 
       it "should save cloned pipeline successfully" do
@@ -772,11 +773,10 @@ describe Admin::PipelinesController do
         current_user_entity_id = 9999
         controller.stub(:current_user_entity_id).and_return(current_user_entity_id)
 
-        selected_pipeline_id = "456"
-
-        expect(@go_config_service).to receive(:updateUserPipelineSelections).with(current_user_entity_id, CaseInsensitiveString.new("new-pip"))
+        expect(@pipeline_selections_service).to receive(:updateUserPipelineSelections).with(current_user_entity_id, CaseInsensitiveString.new("new-pip"))
 
         post :save_clone, :config_md5 => "1234abcd", :pipeline_group => {:group => "group1", :pipeline => {:name => "new-pip"}}, :pipeline_name => @pipeline.name().to_s
+        expect(response.response_code).to be(200)
       end
 
       it "should not update the user's pipeline selections when save clone is not successful" do
@@ -786,7 +786,7 @@ describe Admin::PipelinesController do
 
         post :save_clone, :config_md5 => "1234abcd", :pipeline_group => {:group => "group1", :pipeline => {:name => "new-pip"}}, :pipeline_name => @pipeline.name().to_s
 
-        expect(@go_config_service).to_not receive(:updateUserPipelineSelections)
+        expect(@pipeline_selections_service).to_not receive(:updateUserPipelineSelections)
       end
     end
   end
